test: use test helper functions to instantiate Household
